+++ lib/eve_online/esi/models/model_with_etag.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module EveOnline
+  module ESI
+    module Models
+      module ModelWithEtag
+        def etag
+          options['etag']
+        end
+      end
+    end
+  end
+end