+++ lib/eve_online/esi/universe_stargate.rb
@@ -7,6 +7,8 @@ module EveOnline
     class UniverseStargate < Base
       extend Forwardable
 
+      include ResponseWithEtag
+
       API_ENDPOINT = 'https://esi.evetech.net/v1/universe/stargates/%<stargate_id>s/?datasource=%<datasource>s'
 
       attr_reader :id
@@ -18,10 +20,10 @@ module EveOnline
       end
 
       def_delegators :model, :as_json, :name, :stargate_id, :system_id,
-                     :type_id, :destination, :position
+                     :type_id, :destination, :position, :etag
 
       def model
-        @model ||= Models::Stargate.new(response)
+        @model ||= Models::Stargate.new(response_with_etag)
       end
 
       def scope; end