+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 * Remove `memoist` gem
 * Fix etag support
+* Rename `EveOnline::ESI::Base#no_content?` to `#not_modified?`
+* Add `EveOnline::ESI::Models::ModelWithEtag`
+* Add `EveOnline::ESI::ResponseWithEtag`
+* Add `EveOnline::ESI::Models::Stargate#etag`
+* Add `etag` to `EveOnline::ESI::Models::Stargate#as_json`
+* Add `EveOnline::ESI::UniverseStar#etag`
+* Add `etag` to `EveOnline::ESI::Models::Star#as_json`
 
 ## v0.20.0